Compute IDEA4 indicators and aggregation metrics

Aggregates items from read_idea() to produce IDEA4 indicators, components, dimensions and properties.

compute_idea(data)

Arguments

data
an object of class IDEA_items produced with read_idea()

Value

An object of class "IDEA_data" with three attributes :

metadata

a named list containing the 17 metadata entries about the farm

dataset

a tibble containing the score computed for the 53 indicators, 13 components and 3 dimensions

nodes

a list of tibbles, one per property plus a global one, which all describe the qualitative evaluation obtained for each leaf/node and for the final property.

Details

This function is designed to compute IDEA scores for the dimensions and properties approaches. A copy of the decision rules used for the properties approach can be locally exported as an excel file with :

IDEATools::show_decision_rules()

Further information about decision rules can be found in this vignette :

vignette("decision_rules", package = "IDEATools")

Examples

library(IDEATools)
path <- system.file("idea_example.json", package = "IDEATools")
my_data <- read_idea(path)
computed_data <- compute_idea(my_data)